Separate out exam related models from content app.
