+++ kolibri/content/api.py
@@ -4,10 +4,6 @@ from random import sample
 from django.core.cache import cache
 from django.db.models import Q
 from django.db.models.aggregates import Count
-from django.db.models.query import F
-from django.shortcuts import get_object_or_404
-from kolibri.auth.api import KolibriAuthPermissions, KolibriAuthPermissionsFilter
-from kolibri.auth.filters import HierarchyRelationsFilter
 from kolibri.content import models, serializers
 from kolibri.content.content_db_router import get_active_content_database
 from kolibri.logger.models import ContentSessionLog, ContentSummaryLog
@@ -220,48 +216,3 @@ class FileViewset(viewsets.ModelViewSet):
 
     def get_queryset(self):
         return models.File.objects.all()
-
-class ExamFilter(filters.FilterSet):
-
-    class Meta:
-        model = models.Exam
-        fields = ['collection', ]
-
-class ExamViewset(viewsets.ModelViewSet):
-    serializer_class = serializers.ExamSerializer
-    pagination_class = OptionalPageNumberPagination
-    permissions_classes = (KolibriAuthPermissions,)
-    filter_backends = (KolibriAuthPermissionsFilter, filters.DjangoFilterBackend)
-    filter_class = ExamFilter
-
-    def get_queryset(self):
-        return models.Exam.objects.all()
-
-
-class ExamAssignmentViewset(viewsets.ModelViewSet):
-    serializer_class = serializers.ExamAssignmentSerializer
-    pagination_class = OptionalPageNumberPagination
-    permissions_classes = (KolibriAuthPermissions,)
-    filter_backends = (KolibriAuthPermissionsFilter,)
-
-    def get_queryset(self):
-        return models.ExamAssignment.objects.all()
-
-
-class UserExamViewset(viewsets.ModelViewSet):
-    serializer_class = serializers.UserExamSerializer
-    pagination_class = OptionalPageNumberPagination
-    permissions_classes = (KolibriAuthPermissions,)
-    filter_backends = (KolibriAuthPermissionsFilter,)
-
-    def get_queryset(self):
-        return models.ExamAssignment.objects.all()
-
-    def retrieve(self, request, pk=None, **kwargs):
-        exam = get_object_or_404(models.Exam.objects.all(), id=pk)
-        assignment = HierarchyRelationsFilter(exam.assignments.get_queryset()).filter_by_hierarchy(
-            target_user=request.user,
-            ancestor_collection=F('collection'),
-        ).first()
-        serializer = serializers.UserExamSerializer(assignment, context={'request': request})
-        return Response(serializer.data)
