+++ kolibri/content/api_urls.py
@@ -1,17 +1,14 @@
 from django.conf.urls import include, url
 from rest_framework import routers
 
-from .api import ChannelMetadataCacheViewSet, ContentNodeViewset, ExamAssignmentViewset, ExamViewset, FileViewset, UserExamViewset
+from .api import ChannelMetadataCacheViewSet, ContentNodeViewset, FileViewset
 
 router = routers.SimpleRouter()
 router.register('content', ChannelMetadataCacheViewSet, base_name="channel")
-router.register(r'exam', ExamViewset, base_name='exam')
-router.register(r'examassignment', ExamAssignmentViewset, base_name='examassignment')
 
 content_router = routers.SimpleRouter()
 content_router.register(r'contentnode', ContentNodeViewset, base_name='contentnode')
 content_router.register(r'file', FileViewset, base_name='file')
-content_router.register(r'userexam', UserExamViewset, base_name='userexam')
 
 urlpatterns = [
     url(r'^', include(router.urls)),
